Fixed Makefile .ml compilation dependencies
I had hoped that by properly ordering the ${SOURCES} I could avoid the
odd dependency issues with OCaml. While it works for a 'make clean all',
it does not work for partial recompilation when one of the source files
had its public signature changed. In that case all files depending on
the changed file will have to be recompiled as well. :-(

While I'm glad there's ocamldep, it doesn't integrate all that well with
makefiles, so I decided to manage the dependencies on the git repo, and
added a 'make depend' target to keep this file up-to-date. (Won't have
to be done very often, luckily)