Improved ui_logwindow_draw() and fixed several crashes with large input
ui_logwindow_draw() would previously crash if a log entry consisted of
more than 64k characters, or would be displayed on more than 100 lines.
I've got rid of these crashes by simple not displaying any more than 200
lines or 64k bytes.

Also fixed a bug with the history feature to avoid reading back partial
commands from the history file and to avoid truncating a string in the
middle of a UTF-8 sequence (which, due to all functions assuming correct
UTF-8 on all strings, would cause a crash).

Also did some various cleaning-up: declared some buffers as static (so
they wouldn't need to be allocated on the heap - and none of the
functions are re-entrant in the first place) and removed the redundant
UTF8-to-UCS4-and-back conversion in ui_logwindow_draw(). Not sure which
version is faster, but this one uses less memory. :-)