Check for screen updates every 100ms
This makes ncdc appear a bit faster, since you don't have to wait a
second for new stuff to be visible on the screen. On the other hand,
this slightly increases the CPU usage, but that seems to be negligible
when compared to the CPU usage required for the 1-second-full-redraw.