Use a GList for the user list and keep it sorted
The user list is only stored in a GList while the userlist tab is
opened. If the tab is closed (which it usually is, I'd guess), no sorted
list is kept in memory, thus saving CPU cycles and RAM.

The list is entirely owned and managed by the ui_userlist_* functions,
so I also had to add some mechanism for the nmdc functions to notify it
on changes. (The reason for letting the list be managed by the tab is to
make it easier to handle selecting and paging in the future, which will
be a huge pain otherwise.)