Added basic sorting to the userlist + keep list sorted after changes
Unfortunately, I had to store a GSequenceIter pointer in the nmdc_user
structure anyway. There's no (fast) way to get the iter if you can't be
sure that it is ordered correctly. And we can't, since the sharesize may
change and so may its location in the list when it's sorted by
sharesize.