Implemented TTH functions
They aren't used yet, though. I also haven't extensively tested and
measured the implementation, but so far it seems to be doing its job and
is not slow (getting 220MB/s on one core of my E2160).

Intermediate blocks are being kept up to a certain level in order to
comply with the tthl requirements of modern NMDC/ADC protocols.