filelist.c: Store hash data and lastmod info in a gdbm database
Requires the gdbm library, obviously. According to my roadmap, it looks
like this will be the last required dependency. At least for the initial
release.

I really don't like how this file is going to grow quite a bit on large
shares... but I guess that's what you get for following the crowd and
keeping around 10 levels of hashes. :-(