Store keyboard input in a guint64 instead of a struct
And added macros for generating a key and fetching info from it. This
allows using a switch() statement to check for keyboard input, which is
more readable and easier to maintain.

This trick is also a bit more efficient, but something is terribly wrong
with your system if you notice such an insignificant improvement. :-P