fl_local.c: Keep a tth -> file pointer index + cleanup/improvements
This fixes the bug that used hash data could be deleted when a duplicate
file has been removed from the list. Also did some slight restructuring
to easily allow calculation of the total share size, excluding any
duplicates. The index will later on be useful to reply to TTH searches
and other file-by-TTH-lookups (of which there are quite a few, and they
need to be fast).