util.c: Rewrote net_* functions to use GSocket instead of GIOStream
As you can see, using the low-level GSocket functions results neither in
more code or nor in more complexity, while it is a lot more flexible and
efficient. From this I can conclude that the GIOStream API has failed
its purpose.

The real reason for this change is so that it should now be easier to
add efficient binary transfers - which I'll still have to do.