Added a more efficient implementation of net_sendfile()
Tries to use the sendfile() call when possible, but since that call is
known to not always work, there is a fallback using plain old read() and
send().

The sendfile detection in the configure script might fail on other
systems than Linux, since various other systems have a sendfile() call
as well, which may not be the same as the Linux one.