Restructured nmdc->ui user list notifies + improved completion detection
It's hard to cope with all the differences in how the hubs can send the
userlist. Some hubs don't support NoHello or NoGetINFO (opendchub),
while others refuse to give a $NickList. The current code assumes that
the hub *will* send either a $NickList or a $OpList. And if it sends
only an $OpList, it won't do this before sending all $MyINFO's (this is
what VerliHub and PtokaX do).

I know of one major hub for which the above trick won't work: YnHub. It
refuses to send a $NickList and sends the $OpList before $MyINFO's. I
have no idea how to reliably detect when the list is complete in that
case, so you'll have to live with some annoying "x has joined" on
connect. :-(