Version bump
This git repository now represents version "0.1g", that is, the git repo
after 0.1 and before 0.2. This makes it easier to differentiate between
compiles from the repo and from actual releases. I could include a 'git
describe' in the compilation info, but there isn't really an elegant way
to do so. :-(