ui.c: Replaced g_snprintf() + mvaddstr() with mvprintw()
I was sure there was a printf-like ncurses function when I wrote the
stuff, but was too lazy find it and fell back on g_snprintf() and
mvaddstr() instead. Now that I remembered it was mvprintw(), let's just
use that - it's more clean and not restricted by the size of a
fixed-length buffer.