Make sure all shared files have a case-insensitive unique name
This is to make sure we comply with the ADC spec. This affects the
filename argument of CGET and various virtual path arguments to commands
(e.g. /refresh) (which is now case-insensitive), and affects file list
scanning and reading.

While I was changing the filesystem scanning code, I also fixed a bug
with it reporting non-UTF8 messages to ui_m() in some cases.

This change also introduces a bug: If you rename one of your shared
files and change only the case, ncdc will not pick up the new name and
the file can as such not be downloaded by others until you explicitely
remove and re-add it to your share. I'll see if I can fix this in a
minute.

*UPGRADE NOTICE*
Ncdc will now refuse to open file lists that contain duplicate files
names (in a case-insensitive context). This has a nasty side effect: if
your shared two files with the same name but different case in a single
directory, ncdc will throw away all your hashdata and force a complete
refresh on start-up.