ui_util.c:ui_logwindow_load(): Forgot to remove a buggy g_debug()
Besides not being necessary now that the backlog loading code works fine at
the point where the g_debug() line is, it's also buggy in itself, thus
causing a segfault if the log contains formatting codes. It should have been:
  g_debug("%s", l[i]);

Thanks fly out to xmw for finding and reporting it.