commands.c: Allow backlog to be as large as buffer size of the logwindow
I put the restriction to 100 before as file_tail() would probably be
quite slow with more lines than that. But it may be too restrictive for
those who do want to see more lines. So I upped it to LOGWIN_BUF-1,
which is the number of lines that actually fit in the buffer of a log
window. Setting backlog to more than that is simply pointless.