ui_util.c: Wrap log lines on word boundary
This also has improved support for long "words" (e.g. long URLs), which
are still wrapped on character-boundary, but are not indented after
wrapping. This is similar to what irssi does and allows easy copying of
the multiline URL as a single line - which most terminals support.

(But what I'm doing now is still somehow different from irssi, since a
wrapped URL is still not clickable in roxterm... perhaps I need to write
a complete URL in a single *addstr() function and let ncurses do the
wrapping, in order for the terminal to recognize it as a single URL?)