Force a re-open of the log files when receiving SIGUSR1
If you just want to 'rm' or 'rename' the log files, you could already do
this without problems, since ncdc would automatically detect this and
create a new log file. No need for SIGUSR1 here.

However, it may take a while before the unlinked file is actually freed,
or for the last changes of the renamed file to be flushed. This
disallows doing anything with the renamed log files after renaming them.
The following solution is somewhat more reliable:

  rename some.log /old/some.log
  killall -USR1 ncdc # well, `kill -USR1` on the PID is nicer
  sleep 3; # wait for flush to finish (not too reliable...)

  # now do whatever you want with the renamed file, e.g.:
  xz -9 /old/some.log

A more reliable alternative to the 'sleep' command above is to wait for
some.log to be created again. You can be sure that changes to
/old/some.log have been flushed by that time.

Also, this change does not apply to stderr.log since that one is
special.