Perform a clean shutdown on SIGHUP + Improved signal handling
ncdc will now cleanly exit when the terminal in which it runs is killed
or closed.

The signal handling has been changed to only set global variables when
receiving the signals, which are then checked and acted upon in the main
loop. For signals that require immediate attention (SIGHUP, SIGTERM,
SIGUSR1), a special GSource was created that simply checks for these
global variables before and after the poll() in the main event loop.
This ensures "immediate" attention in a synchronous fashion, as poll()
is interrupted when a signal arrives.

I also added a noterm variable, which avoids writing anything to the
terminal when shutting down. I don't know how the write() functions
react after a SIGHUP, after all. This trick does not work when SIGHUP is
received while ncdc is busy in ui_draw(), but oh well...