Started on a rewrite of the network code to the High-level GIO API
This is pretty much the opposite of what I decided in
  2e5ca63194ef0667a739547edf4e60b1ab5c2817
The reason for this change is because the GIO API allows for TLS (and
perhaps also SOCKS) support without having to maintain several
code-paths by myself.

Even with my better knowledge of the GIO API now, the higher-level
abstractions (GDataInputStream and GBufferedOutputStream) which I
attempted to use in earlier versions is, indeed, definitely not suitable
for use in ncdc. There's also a slight problem that the transfer rates
as is currently calculated is inaccurate when TLS or SOCKS is used. GIO
does not provide a more reliable alternative, so we'll have to do with
what we have.

This commit breaks file uploading and handling of incoming connections.
I'll fix that in later commits.