Re-implemented file sending code, in a separate thread this time
The file_* pretty much mimic the behaviour of a _splice() in the GIO
API. Except that my version is more useful.

Note that timeout_last is not updated from the upload thread, instead
the timeout function checks for changes to file_left and updates
timeout_last when necessary. This may result in an error of some 5
seconds, but that is more than enough to prevent an idle timeout while
uploading.

I also figured that my previous method for handling a disconnect and/or
free() of the net structure while an _async() operation was in progress
wasn't good enough: checking whether it returned a CANCELLED error isn't
realiable, since the disconnect/free may also have happened after the
operaiton finished but before the ready callback has been called,
meaning the net struct is freed without the callback knowing this.
Solution: do a net_ref() before performing an async operation.

Something similar applies to the actual I/O stream objects: the GIO
internals don't handle an unref() from another thread when an async
operation is in progress. So these objects have to be _ref()'ed as well.

Seriously, plain old sockets are a thousand times easier. :-/