Added basic manual page for ncdc and ncdc-gen-cert
Based on the man page kaste wrote, which was supposedly based on
WeeChat's man page.

I haven't included anything about the interactive commands or key
bindings yet, still have to figure out how to handle that in a nice way.
So this man page is really quite basic.

A simple man page for ncdc-gen-cert is included under the philosophy of
"everything in /usr/bin/ must have a man page with the same name".