ui_util.c: Improved line wrapping algorithm
Spaces at the beginning of a new line are now removed, and the decision
to char-wrap long words has been changed to handle the case that the
word is longer than the available space on a line, including
indentation. Not counting indentation, long words may now be longer than
cols/2 and avoid getting wrapped (it's cols*3/4 now).