net.c: Don't wrap GTlsConnection around a GTcpWrapperConnection
There's no need for it in the net.c code, since the GTlsConnection
object is created in the net.c code and not some call from GIO.

This also happens to fix a minor bug with uncleanly disconnecting a TLS
connection. Apparently GTcpWrapperConnection doesn't invoke the _close()
method of the wrapped object (or something?).