dl.c: Improved connection state notification (NCO->EXP->IDL)
dl.c now has less knowledge about what happens in hub.c and cc.c.
NCO->EXP is done in dl.c itself, when it requests a connection. EXP->IDL
is a notify from cc.c, when it gets into the IDLE state. *->WAI is
done with the same notify as EXP->IDL, and is performed at various
places in cc.c.

Note that this commit completely breaks the current downloading code.