dl.c: Perform filename conflict resolution after download is complete
This is more reliable, because the contents of the download_dir may be
been changed, thus creating another filename conflict. Actually, such a
conflict would already be handled by GIO when moving, but then the
previously existing file would be renamed, which isn't what we want.

This solution also happens to be significantly faster.