Keep track of parent for each tab, and open that when closing the tab
This is a different take on the previous solution, which tried to
determine a "parent" tab from some rules (i.e. parent tab of a file
browser for a user is the hub tab where that user is online).

This new approach actually tracks from which tab a tab was opened. So if
you now open a users' file list from the user list tab (the 'b' key),
you will go back to the user list again after closing the browse tab -
rather than the hub tab as the previous approach would.

In various cases no parent tab will be assigned to a certain tab, this
holds for e.g. '/search -all', '/open', '/connections', and a few more.

When a tab that is assigned as parent to other tabs is closed, the new
parent for those tabs will be the parent of the closing tab. Now that
explanation sounds rather annoying to understand; the code is easier -
it's in ui.c:ui_tab_remove().