Abstracted creation of fl_list structures + more efficient allocating
I now use the slice allocator for fl_list objects, since I can always
determine the size of one quite easily. The slice allocator is slightly
more efficient in terms of memory use.

I also use a more tight calculation for the size of the final
structures, sizeof() would allocate more space than necessary since it
pads the structure to a multiple of 8 bytes, which isn't very useful if
you later add a strlen() to that. The current method determines the
actual size of the struct up to the name[] array, and adds strlen()+1
more bytes to store the name itself. This trick *should* be fairly
portable, but I guess it's a matter of testing to prove that...