ui.c: Use (uid -> tab) hash table to keep track of MSG tabs
This replaces a linear search through the tab list with a hash table
lookup. This isn't a very significant improvement at the moment since
this lookup is rarely used, but is definitely necessary when
implementing proper user status notifications for MSG tabs - in which
case this lookup is done every time a user joins or quits the hub.