fl_util.c: Store data for local files at the end of fl_list structs
This way, the memory for local data is only allocated for the local file
list, not for others. This saves a bit of memory.

The main reason for this change is that it allows me to add more fields
to the local file list without having to worry about increasing the
memory requirements for other people's file list. It's a bit of an ugly
hack, but it appears to work fine. I also replaced the fl_list_size()
function with a less ugly macro.

Note that fl_list_copy() still makes a perfect copy of the file list,
thus including the `islocal' value and the associated local data. This
isn't too optimal when creating a copy for browsing purposes, where this
local data isn't used. Not that much of a problem, but oh well.