db.c/dl.c: Add dl_user rows to the database
I believe that was the last one, in terms of keeping the database
synchronized with what actually happens. What's left is making sure that
the hash data is read back from the database again to check it against
the file. This is going to require some rewriting to turn db.c into a
full message passing-based implementation, I'm afraid. :-(