db.c: Don't queue queries for batch processing in background thread
This is a partial rewrite of the background thread that communicates
with the SQLite database. Previously it would try to make sure that a
transaction is as "short" (in time) as possible, while making sure to
utilize a single transaction as much as possible within that time.

This new solution doesn't: It keeps a transaction open and processes
incoming query immediately, managing transaction starting/committing
in-between.

This has one main advantage: Queries that want to have a result back
from the database don't need to use DBF_SINGLE or DBF_LAST anymore, thus
avoiding an unecessary flush of the current transaction. This
significantly improves performance and responsiveness of SELECT queries
or INSERTs that wish to have a lastid. In particular, hash checking of
download data should have less impact on responsiveness, and saving
hash data of shared files should have less impact on hashing speed. (I
measured the last one, and that definitely improved)

A disadvantage (and the reason why I initially didn't do this), is that
ncdc is now slightly less resilient to errors that may occur when the
database is also modified by other processes. But first of all this
isn't very common, and secondly I doubt all that much can go wrong -
SQLite isn't that badly written.

I also increased the flush timeout from 1 to 5 seconds, thus allowing
even more queries to be batched in a single transaction. (...and causing
more data loss in the event of an error, but oh well.)

tl;dr: ncdc is now faster and more responsive.