net.c/cc.c: Use normal integers for the transfer size of downloads
I already did this with uploads, in order to allow atomic integer
operations to be used for accessing those in multiple threads. As with
uploads, this has the limitation that files can only be transferred in
chunks of max. 2GiB. I don't see this as much of a problem, though.