cc.c: Reply with same byte count in ADCSND even when chunk >=2GB
This isn't much of a problem, except that the download halts at 2GB,
the connection will be closed 30 seconds later due to inactivity, and
the downloading client will eventually continue with a new chunk. A
minor inconvenience, but not as bad as DC++ and derivatives refusing to
download anything because the byte count doesn't match.